Refactor ldp.container.post
