+++ src/middleware/packages/ldp/services/resource/index.js
@@ -1,10 +1,11 @@
 const getAction = require('./actions/get');
-const postAction = require('./actions/post');
+const createAction = require('./actions/create');
 const patchAction = require('./actions/patch');
 const putAction = require('./actions/put');
 const deleteAction = require('./actions/delete');
 const existAction = require('./actions/exist');
 const generateIdAction = require('./actions/generateId');
+const uploadAction = require('./actions/upload');
 const headAction = require('./actions/head');
 const methods = require('./methods');
 
@@ -19,11 +20,11 @@ module.exports = {
   actions: {
     exist: existAction,
     generateId: generateIdAction,
+    create: createAction,
+    upload: uploadAction,
     // Actions accessible through the API
     api_get: getAction.api,
     get: getAction.action,
-    api_post: postAction.api,
-    post: postAction.action,
     api_patch: patchAction.api,
     patch: patchAction.action,
     api_delete: deleteAction.api,
